requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
google-api-core==1.11.1
google-auth==1.6.3
google-cloud-core==1.0.1
google-cloud-translate==1.5.0
googleapis-common-protos==1.6.0
grpcio==1.21.1
idna==2.8
pbr==5.2.1
pipenv-to-requirements==0.7.1
pipenv==2018.11.26
protobuf==3.8.0
pyasn1-modules==0.2.5
pyasn1==0.4.5
pydub==0.23.1
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.3
virtualenv-clone==0.5.3
virtualenv==16.6.0
